gateway/app/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
redis
pika
prometheus-client
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi